# Conflicts:
#	Build/Viewer/source/main_gl.cpp

+++ Common/Animator/MersenneTwister.h
@@ -105,7 +105,7 @@ public:
 	}
 	// generates a random number on [0,0xffffffff]-interval
 	unsigned long genrand_uint32() {
-		unsigned long aulMag01[2] = { 0x0UL, m_ulMATRIX_A };
+		unsigned long aulMag01[2] = { 0x0UL, (unsigned long)m_ulMATRIX_A };
 		// mag01[x] = x * MATRIX_A  for x=0,1
 		if (m_iMTI >= m_iN) {
 			// generate N words at one time
